requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
yfinance
pandas
numpy
python-dotenv
apscheduler